Embed vault metadata in FinderItems to eliminate lookup overhead
Fetches vault metadata once in fetchFinderItems queryFn and embeds it directly into each FinderItem, removing the need for Map lookups during filtering. This eliminates re-renders caused by vaultsMetaMap changes and improves UI responsiveness.
